[int C] Avoid comparing signed/unsigned.
* src/rcsedit.c (deletelines): Take `unsigned long' (both args).
Also, use `unsigned long' for local var.
(insertline): Change type of first arg to `unsigned long'.
(resolve_symlink): Use `ssize_t' for local var.
* src/rcsfnms.c (pairnames): Add cast to `size_t' for `strlen' rv.
Also, transform `baselen' comparison algebraically.
* src/rcslex.c (fd2RILE): Use `off_t' for local var.
* src/rcsutil.c (write_stderr): Use `int' for local var.

+++ src/rcslex.c
@@ -857,7 +857,7 @@ fd2RILE (int fd, char const *name, char const *type,
         static RILE rilebuf[RILES];
 
         register RILE *f;
-        size_t s = status->st_size;
+        off_t s = status->st_size;
 
         if (s != status->st_size)
           faterror ("%s: too large", name);
